+++ packages/runtime/plugin-runtime/src/core/context/serverPayload/index.server.tsx
@@ -1,5 +1,34 @@
 import { storage } from '@modern-js/runtime-utils/node';
-import type { ServerPayload } from '../index';
+import type { RouterState } from '@modern-js/runtime-utils/remix-router';
+
+export type PayloadRoute = {
+  clientAction?: any;
+  clientLoader?: any;
+  element?: React.ReactNode;
+  errorElement?: React.ReactNode;
+  handle?: any;
+  hasAction: boolean;
+  hasErrorBoundary: boolean;
+  hasLoader: boolean;
+  id: string;
+  index?: boolean;
+  params: Record<string, string>;
+  parentId?: string;
+  path?: string;
+  pathname: string;
+  pathnameBase: string;
+  shouldRevalidate?: any;
+};
+
+export type ServerPayload = {
+  type: 'render';
+  actionData: RouterState['actionData'];
+  errors: RouterState['errors'];
+  loaderData: RouterState['loaderData'];
+  location: RouterState['location'];
+  routes: PayloadRoute[];
+  originalRoutes?: PayloadRoute[];
+};
 
 export const getServerPayload = (): ServerPayload | undefined => {
   const context = storage.useContext() as any;
