+++ packages/runtime/plugin-runtime/src/core/context/serverPayload/index.tsx
@@ -1,4 +1,5 @@
-import type { ServerPayload } from '../index';
+import type { PayloadRoute, ServerPayload } from './index.server';
+export type { ServerPayload, PayloadRoute };
 
 export const getServerPayload = (): ServerPayload | undefined => {
   return undefined;
